// interfaces/menu.interfaces.ts
import { TopLevelCategory } from "./page.interface";


export interface PageItem{ 
    alias: string;
    title: string; 
    _id: string; 
    category: string; 
}

export interface MenuItem{
    _id:{
        secondCategory: string; 
    };
    isOpend?: boolean;
    pages: PageItem[];
}

export interface FirstLevelMenuItem{
    route: string;
    name: string;
    icon: JSX.Element;
    id: TopLevelCategory;
}